stage 2
Stage 2/3: Upper or lower
Description
Let's work with the text in the textarea element. In this stage, we will add some handlers for buttons.

Your application should be able to convert a text into the upper case, lower case, proper case, and sentence case.

The upper case is the case when each letter is in the upper case. For example, THIS IS THE UPPER CASE TEXT.
The lower case is the case when each letter is in the lower case. For example, this is the lower case text.
The proper case is the case when each word starts with the upper case, and the rest of the word is in the lower. For example, This Is The Proper Case Text.
The sentence case is the case when each sentence starts with an upper case letter, and the rest of the sentence is in the lower case. In example, This is the sentence case. Only the first word of the sentence starts with an upper case letter.
This stage requires knowledge of various string methods. You can take a look at the W3 Schools Tutorial. Consider using the join() method.

Objectives
Your web page should contain all elements from the previous stage.

Add the following event handlers to the buttons:

If the upper-case button is pressed, convert the text in textarea to the upper case;
If the lower-case button is pressed, convert the text in textarea to the lower case;
If the proper-case button is pressed, convert the text in textarea to the proper case;
If the sentence-case button is pressed, convert the text in textarea to the sentence case.