stage 3/3
Stage 3/3: Save me
Description
You can convert the text into different cases, excellent! Now let's add the ability to save the changed text as a .txt file!

Objectives
Add one more button to the page. Assign the save-text-file id to it. Add another event handler for the button.

When the button is clicked, you should get the text in the textarea element and generate the text.txt file. The resulting file should contain the text from the textarea element.